+++ include/arch/arc/arch.h
@@ -28,6 +28,7 @@
 #include <arch/arc/v2/arcv2_irq_unit.h>
 #include <arch/arc/v2/asm_inline.h>
 #include <arch/arc/v2/addr_types.h>
+#include "v2/sys_io.h"
 #endif
 
 #ifdef __cplusplus
