+++ include/arch/common/sys_io.h
@@ -1,25 +1,35 @@
 /*
- * Copyright (c) 2016 Cadence Design Systems, Inc.
+ * Copyright (c) 2015, Wind River Systems, Inc.
+ * Copyright (c) 2017, Oticon A/S
+ *
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef ZEPHYR_INCLUDE_ARCH_XTENSA_SYS_IO_H_
-#define ZEPHYR_INCLUDE_ARCH_XTENSA_SYS_IO_H_
+/* "Arch" bit manipulation functions in non-arch-specific C code (uses some
+ * gcc builtins)
+ */
 
-#if !defined(_ASMLANGUAGE)
+#ifndef ZEPHYR_INCLUDE_ARCH_COMMON_SYS_IO_H_
+#define ZEPHYR_INCLUDE_ARCH_COMMON_SYS_IO_H_
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#ifndef _ASMLANGUAGE
+
+#include <zephyr/types.h>
 #include <sys_io.h>
 
-/* Memory mapped registers I/O functions */
 
-static ALWAYS_INLINE u32_t sys_read32(mem_addr_t addr)
+static ALWAYS_INLINE u8_t sys_read8(mem_addr_t addr)
 {
-	return *(volatile u32_t *)addr;
+	return *(volatile u8_t *)addr;
 }
 
-static ALWAYS_INLINE void sys_write32(u32_t data, mem_addr_t addr)
+static ALWAYS_INLINE void sys_write8(u8_t data, mem_addr_t addr)
 {
-	*(volatile u32_t *)addr = data;
+	*(volatile u8_t *)addr = data;
 }
 
 static ALWAYS_INLINE u16_t sys_read16(mem_addr_t addr)
@@ -32,15 +42,14 @@ static ALWAYS_INLINE void sys_write16(u16_t data, mem_addr_t addr)
 	*(volatile u16_t *)addr = data;
 }
 
-
-static ALWAYS_INLINE u8_t sys_read8(mem_addr_t addr)
+static ALWAYS_INLINE u32_t sys_read32(mem_addr_t addr)
 {
-	return *(volatile u8_t *)addr;
+	return *(volatile u32_t *)addr;
 }
 
-static ALWAYS_INLINE void sys_write8(u8_t data, mem_addr_t addr)
+static ALWAYS_INLINE void sys_write32(u32_t data, mem_addr_t addr)
 {
-	*(volatile u8_t *)addr = data;
+	*(volatile u32_t *)addr = data;
 }
 
 /* Memory bit manipulation functions */
@@ -61,24 +70,9 @@ static ALWAYS_INLINE void sys_clear_bit(mem_addr_t addr, unsigned int bit)
 
 static ALWAYS_INLINE int sys_test_bit(mem_addr_t addr, unsigned int bit)
 {
-	int temp = *(volatile int *)addr;
+	u32_t temp = *(volatile u32_t *)addr;
 
-	return (int)(temp & (1 << bit));
-}
-
-static ALWAYS_INLINE int sys_test_and_set_bit(mem_addr_t addr, unsigned int bit)
-{
-	int retval = (*(volatile int *)addr) & (1 << bit);
-	*(volatile int *)addr = (*(volatile int *)addr) | (1 << bit);
-	return retval;
-}
-
-static ALWAYS_INLINE
-	int sys_test_and_clear_bit(mem_addr_t addr, unsigned int bit)
-{
-	int retval = (*(volatile int *)addr) & (1 << bit);
-	*(volatile int *)addr = (*(volatile int *)addr) & ~(1 << bit);
-	return retval;
+	return temp & (1 << bit);
 }
 
 static ALWAYS_INLINE
@@ -102,6 +96,27 @@ static ALWAYS_INLINE
 	return sys_test_bit(addr + ((bit >> 5) << 2), bit & 0x1F);
 }
 
+static ALWAYS_INLINE
+	int sys_test_and_set_bit(mem_addr_t addr, unsigned int bit)
+{
+	int ret;
+
+	ret = sys_test_bit(addr, bit);
+	sys_set_bit(addr, bit);
+
+	return ret;
+}
+
+static ALWAYS_INLINE
+	int sys_test_and_clear_bit(mem_addr_t addr, unsigned int bit)
+{
+	int ret;
+
+	ret = sys_test_bit(addr, bit);
+	sys_clear_bit(addr, bit);
+
+	return ret;
+}
 
 static ALWAYS_INLINE
 	int sys_bitfield_test_and_set_bit(mem_addr_t addr, unsigned int bit)
@@ -124,7 +139,11 @@ static ALWAYS_INLINE
 
 	return ret;
 }
-#endif /* !_ASMLANGUAGE */
 
+#endif
 
-#endif /* ZEPHYR_INCLUDE_ARCH_XTENSA_SYS_IO_H_ */
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* ZEPHYR_INCLUDE_ARCH_COMMON_SYS_IO_H_ */
