+++ include/arch/nios2/nios2.h
@@ -53,13 +53,13 @@ extern "C"
 
 #include <zephyr/types.h>
 #include <arch/cpu.h>
-#include <sys_io.h>
+#include <arch/common/sys_io.h>
 
 /*
  * Functions for accessing select Nios II general-purpose registers.
  */
 
-/* ET (Exception Temporary) register */ 
+/* ET (Exception Temporary) register */
 static inline u32_t _nios2_read_et(void)
 {
 	u32_t et;
